requirements.txt
notify_py==0.3.43
psutil==7.0.0
python-dotenv==1.1.1
twilio==9.6.5
notify-py==0.3.43
